some data cleansing and added user confirmation for deletion

+++ public/assets/js/burgers.js
@@ -1,20 +1,25 @@
 $(function() {
-    var liToEat = $(".liToEat");
+    function checkEmpty(){
+        let liToEat = $(".liToEat");
+        let liAte = $(".liAte");
 
-    if(liToEat.length) {$(".left-div ul h3").text("Ready to Eat..")}
-    else {$(".left-div ul h3").text("Add some burgers to Enjoy...")}
-    
-    var liAte = $(".liAte");
+        if(liToEat.length) {$(".left-div ul h3").text("Ready to Eat..")}
+        else {$(".left-div ul h3").text("Add some burgers to Enjoy...")}
+        
+        if(liAte.length) {$(".right-div ul h3").text("Woww..You are doing great!")}
+        else {$(".right-div ul h3").text("Hungry!!!")}
+    }
 
-    if(liAte.length) {$(".right-div ul h3").text("Woww..You are doing great!")}
-    else {$(".right-div ul h3").text("Hungry!!!")}
-    
     $(".create-form").on("submit", function(event){
         event.preventDefault()
         
         if($("#burgName").val()){
-            var newBurger = {
-                burgerName: $("#burgName").val().trim()
+            let re = /(\b[a-z](?!\s))/g;
+            let burgName = $("#burgName").val().trim()
+            burgName = burgName.replace(re, (x) => {return x.toUpperCase()})
+            
+            let newBurger = {
+                burgerName: burgName
             }
             
             $.ajax("api/burgers", {
@@ -30,10 +35,10 @@ $(function() {
     })
 
     $(".eatNoeat").on("click", function(event){
-        var id = $(this).data("id")
-        var devoured = !($(this).data("devour"))
+        let id = $(this).data("id")
+        let devoured = !($(this).data("devour"))
 
-        var newDevoured = {
+        let newDevoured = {
             devouredVal: devoured
         }
         
@@ -49,15 +54,20 @@ $(function() {
     })
 
     $(".delete").on("click", function(event){
-        var id = $(this).data("id")
+        let id = $(this).data("id")
+        let reply = confirm("Are you sure about removing " + $(this).data("name") + " from the list and Database?");
         
-        $.ajax("api/burgers/" + id,{
-            type: "DELETE",
-        }).then(
-            function(){
-                console.log("deleted the burger ", id)
-                location.reload()
-            }
-        )
+        if(reply){
+            $.ajax("api/burgers/" + id,{
+                type: "DELETE",
+            }).then(
+                function(){
+                    console.log("deleted the burger ", id)
+                    location.reload()
+                }
+            )
+        }
     })
+    checkEmpty()
 })
+
